Add PyPI-independent verification; fix WebGL2 shader precision bug
The render smoke test caught a real bug: WebGL2 rejects a program whose shared
uniforms (u_width) differ in precision between stages — vertex floats default to
highp, fragment shaders declared mediump. Fixed both fragment shaders to highp.

Two new stdlib-only gates that run without numpy/PyPI (and first in CI):
- scripts/abi_smoke.py: 12 checks over the Rust C ABI via ctypes + array —
  precision, M4 spikes, zone-map NaN handling, invalid-arg sentinels
- scripts/render_smoke_nonumpy.py: hand-builds a wire payload and drives
  headless Chromium, reads back lit pixels + DOM chrome nodes

Verified locally: cargo test 8/8, ABI smoke 12/12, render smoke lit+chrome OK.

+++ js/src/fastcharts.js
@@ -206,7 +206,9 @@ void main() {
 }`;
 
 const POINT_FS = `#version 300 es
-precision mediump float;
+// highp (not mediump): vertex-stage floats default to highp, and WebGL2 rejects
+// a program whose shared uniforms differ in precision between stages.
+precision highp float;
 uniform vec4 u_color;
 out vec4 outColor;
 void main() {
@@ -243,7 +245,8 @@ void main() {
 }`;
 
 const LINE_FS = `#version 300 es
-precision mediump float;
+// highp: u_width is shared with the vertex stage (highp there); precisions must match.
+precision highp float;
 uniform vec4 u_color; uniform float u_width;
 in float v_off;
 out vec4 outColor;
